+++ src/Traits/WithCollapsingColumns.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits;
+
+use Rappasoft\LaravelLivewireTables\Traits\Configuration\CollapsingColumnConfiguration;
+use Rappasoft\LaravelLivewireTables\Traits\Helpers\CollapsingColumnHelpers;
+
+trait WithCollapsingColumns
+{
+    use CollapsingColumnConfiguration;
+    use CollapsingColumnHelpers;
+    
+    protected bool $collapsingColumnsStatus = true;
+
+}
